Code Reorganization and Resource Producers
I've made some significant improvements to code organization, and
simplified a lot of the logic for interactions between buildings and
players. Player selection also correctly activates or deactivates
players based on whether they are in the game or not.

I have also created the script that controls the most basic type of
building: the resource producer. Resource producers constantly produce
resources of a single type. Upon a player's action, they sell all their
stockpiled resources to the player at a fixed cost (the owner also gains
1 of that resource, but I haven't implemented that yet).