Made demo of fishery updatable graphics
I created a demonstration of how resource buildings will display their
current offer using the fishery. The building displays the resource
offered, the price (in the lower corner), and a number indicating how
many of that resource is available. Pressing A causes the player to
purchase the indicated amount of resources for the indicated price.